[1주차] countBall, isBall, currentBall 메서드 및 테스트 코드 구현
