+++ src/main/java/baseball/core/Core.java
@@ -6,16 +6,47 @@ public class Core {
     public int countStrike(List<Integer> computerNumbers, List<Integer> playerNumbers) {
         System.out.println("스트라이크 확인 합시다!!");
         int size = playerNumbers.size();
-        int count = 0;
+        int strike = 0;
         for (int index = 0; index < size; index++) {
             if (isStrike(computerNumbers, playerNumbers, index)) {
-                count++;
+                strike++;
             }
         }
-        return count;
+        return strike;
     }
 
     private boolean isStrike(List<Integer> computerNumbers, List<Integer> playerNumbers, int index) {
         return computerNumbers.get(index).equals(playerNumbers.get(index));
+    }
+
+    public int countBall(List<Integer> computerNumbers, List<Integer> playerNumbers) {
+        System.out.println("볼 확인 합시다!");
+        int size = playerNumbers.size();
+        int ball = 0;
+        for (int index = 0; index < size; index++) {
+            if (isBall(computerNumbers, playerNumbers, index)) {
+                ball++;
+            }
+        }
+        return ball;
+    }
+
+    private boolean isBall(List<Integer> computerNumbers, List<Integer> playerNumbers, int index) {
+        boolean result = false;
+
+        int size = computerNumbers.size();
+        for (int diff = 0; diff < size; diff++) {
+            if (index != diff) {
+                result = currentBall(playerNumbers.get(index), computerNumbers.get(diff));
+            }
+            if (result) { // 볼이 발견되면 반복 멈춤!
+                break;
+            }
+        }
+        return result;
+    }
+
+    private boolean currentBall(int computerNumber, int playerNumber) {
+        return computerNumber == playerNumber;
     }
 }
